requirements.txt
numpy==1.19.4
scipy==1.4.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
seaborn==0.11.0
sklearn==0.0
sklearn-pandas==1.8.0
tweepy==3.6.0
#wordcloud==1.5.0
urllib3==1.24.3
nltk==3.2.5
matplotlib==3.2.2
joblib==1.0.0
textblob
gensim
pyLDAvis